Updates subscriptions
As records are consumed, we now update the offsets we observe per topic/partition. If we then must re-subscribe given a loss of connectivity, we use the updated offsets as our starting point. If we did not do this then we would start consuming from our original start point each time.

The subscriptions argument has also had its optionality removed. There's no point in subscribing with no subscriptions!

Finally, if we do re-connect given no more chunks then we reset the delayer so that we can attempt to resume asap. I've noticed that reqwest (or hyper) times out the request after a little - which is probably good from a proxying/connection perspective. Thus, it is a normal condition to receive no chunk and then have to re-connect.

The tests have been updated to ensure that the consumer idle behaviour works properly.