Ensure the active file is opened
We must open the active file prior to obtaining its size. Otherwise, we get 0 back as a file size and the compactor won't kick in when restarting again and again e.g. in test scenarios, but there could also be real scenarios. Events tend to get written again with restarts, so a continual restart situation should indeed kick the compactor off when needed.