Additional check on secret len and sync
I'm thinking it is possible to be able to open a file but then not receive bytes on a read given that the other file hasn't yet been flushed. We now have an additional check prior to splitting the string for the salt, and synchronise the file system to flush everything.