Introduces key state for compaction
Key state for compaction means that we can hold state specifically so that keys can be a string that maps to a u64. Prior to this change, keys could only ever be derived from the record key field or by looking further into the ConsumerRecord's value field. Sometimes though, keys simple cannot be converted to a number e.g. when the identity of a record is represented by a string. Given key state, we can now hold that mapping between one type and a u64.