Improve subscription polling efficiency
Subscriptions would wake up every 100ms or so and then read through the log to publish anything written since the last offset. This could cause a lot of record reading and deserialisation without yielding anything new. Profiling of another project using streambed-logged highlighted this as a major area of CPU consumption.

The change here is to quickly check if the last log file has been modified since we last looked. If it hasn't, go to sleep again. This saves a large amount of processing.